requirements.txt
flask
flask-cors
python-dotenv
requests
sentence-transformers
faiss-cpu
numpy
tqdm
torch
